test: Separate crop tests
